+++ packages/cli/CHANGELOG.md
@@ -10,6 +10,9 @@
 - Updated `moon ci` to include a summary of all failed actions.
 - Updated `moon run` to compare against the previous commit when running on the default branch and
   using `--remote`.
+- Updated `rust.bins` in `.moon/toolchain.yml` to support an object for each bin entry.
+  - Can denote bins as CI or local only.
+  - Can force install bins.
 - Updated the run report to include stderr/stdout for all attempts.
 
 #### 🐞 Fixes